// src/Interfaces/IChamps.ts
export interface IChamps {
  ID_Champs?: number;
  culture: string;
  variete: string;
  supTotal: number;
  supInfecte: number;
  dateSemi: string;
  irrigation: string;
  engrai: string;
  stadeCroissance: string;
  santeGle: string;
  nomRavageur: string;
  localisation: string;
  ID_Producteur: number;
}